화장실, 쓰레기통 기준으로 리뷰 전체 조회

+++ src/main/java/dev/changuii/project/dto/response/ResponseReviewDTO.java
@@ -16,12 +16,11 @@ public class ResponseReviewDTO {
     private Long id;
     private String nickname;
     private String password;
-    private int rate;
+    private Double rate;
     private List<String> tag;
     private String content;
     private ToiletDTO toiletDTO;
     private GarbageBinDTO garbageBinDTO;
     private LocalDateTime writeDate;
-
 
 }
